+++ R/vars-rename.R
@@ -29,13 +29,12 @@ vars_rename <- function(.vars, ..., .strict = TRUE) {
     }
   }
 
-  # Remove existing variables
-  if (length(old_vars)) {
-    dups_idx <- match(names(old_vars), .vars, nomatch = 0L)
-    if (!vec_index_is_empty(dups_idx)) {
-      .vars <- .vars[-dups_idx]
-    }
-  }
+  rename_check(
+    to = new_vars,
+    vars = setdiff(.vars, old_vars),
+    orig = .vars,
+    incl = set_names(match(old_vars, .vars), names(old_vars))
+  )
 
   select <- set_names(.vars, .vars)
   renamed_idx <- match(old_vars, .vars)